# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "IWB Submission API",
        "contact": {
            "name": "Liam Griffiths",
            "email": "me@liam-griffiths.co.uk"
        },
        "version": "1.0"
    },
    "paths": {
        "/api": {
            "get": {
                "description": "Will respond I'm Alive! if the API is up.",
                "produces": [
                    "application/json"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/api-docs": {
            "get": {
                "description": "Swagger docs.",
                "produces": [
                    "application/json"
                ],
                "summary": "Swagger docs.",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "description": "Will respond I'm Alive! if the API is up.",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete product",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "description": "Will respond I'm Alive! if the API is up.",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete product",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "description": "Will respond I'm Alive! if the API is up.",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete product",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "description": "Will respond I'm Alive! if the API is up.",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete product",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}